requirements/_base.txt
django==1.10.4
psycopg2==2.6.2
django-environ==0.4.1
django-markdown-app==0.9.0
whitenoise==3.2.2
